% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpls_package_practice.R
\name{mse}
\alias{mse}
\title{Get Root Mean Squared Error of Matrix}
\usage{
mse(est.val, true.val)
}
\arguments{
\item{est.val}{The matrix or vector of estimated values}

\item{true.val}{The matrix or vector of true values}
}
\value{
The RMSE value
}
\description{
Generate RMSEs for a given true and estimated matrix or vector
}
