% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpls_package_practice.R
\name{pred_y}
\alias{pred_y}
\title{Find Predicted Fitted Values}
\usage{
pred_y(testX, testY, coeffs)
}
\arguments{
\item{testX}{The matrix of predictor variables}

\item{testY}{The matrix of response variables}

\item{coeffs}{The list of coefficient matrices to generate fitted values}
}
\value{
A list of fitted values.
}
\description{
Generate a list of Y matrices for each number of principle components
}
\examples{
data(nutrimouse);
simpls_B <- simpls(nutrimouse$genotype, nutrimouse$lipid, cov(nutrimouse$genotype))$coefficients;
fitted <- pred_y(nutrimouse$genotype, nutrimouse$lipid, simpls_B);
}
