% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpls_package_practice.R
\name{simpls}
\alias{simpls}
\title{Proposed SIMPLS Method}
\usage{
simpls(X, Y, dd)
}
\arguments{
\item{X}{The matrix of predictor variables}

\item{Y}{The matrix of response variables}

\item{dd}{The covariance matrix}
}
\value{
A list of the following: the coefficient matrix from the maximum number of principle components
a scores matrix, leverages, the x variance and y variance matrices, the number of principle 
components used, a list of coefficient matrices by number of principle components used, the P
matrix, R matrix, V matrix, and a list of fitted values by number of principle components used.
}
\description{
Find matrix of coefficients and fitted values from given data.
}
\examples{
data(nutrimouse);
simpls_B <- simpls(nutrimouse$genotype, nutrimouse$lipid, cov(nutrimouse$genotype))$coefficients;
fitted <- pred_y(nutrimouse$genotype, nutrimouse$lipid, simpls_B);
}
